more linting

+++ pkg/consensus/pow.go
@@ -22,7 +22,7 @@ func (pow *ProofOfWork) CalculateBlockHash(b *kernel.Block) ([]byte, uint, error
 	var hash []byte
 
 	hashDifficulty := big.NewInt(1)
-	hashDifficulty.Lsh(hashDifficulty, uint(256-b.Target))
+	hashDifficulty.Lsh(hashDifficulty, 256-b.Target)
 
 	maxNonce := ^uint(0)
 	nonce := uint(0)
